Refactor database.py -> databases/*.py, each db gets a file.
