+++ data_diff/diff_tables.py
@@ -12,8 +12,8 @@ from concurrent.futures import ThreadPoolExecutor
 from runtype import dataclass
 
 from .sql import Select, Checksum, Compare, DbPath, DbKey, DbTime, Count, TableName, Time, Min, Max
-from .database import Database
-from .database_types import NumericType, PrecisionType, UnknownColType, Schema, Schema_CaseInsensitive, Schema_CaseSensitive
+from .databases.base import Database
+from .databases.database_types import NumericType, PrecisionType, UnknownColType, Schema, Schema_CaseInsensitive, Schema_CaseSensitive
 
 logger = logging.getLogger("diff_tables")
 
@@ -32,10 +32,6 @@ def safezip(*args):
 def split_space(start, end, count):
     size = end - start
     return list(range(start, end, (size + 1) // (count + 1)))[1 : count + 1]
-
-
-def parse_table_name(t):
-    return tuple(t.split("."))
